Add files via upload
You will see that the training takes a lot of time because of the amount of images fed to the network, if you write a simple print counter in the loop you will see that it takes about 15 seconds for a picture to be propagated. You can play with the if statement on line 136 to get pictures for each epoch, I only managed to train 3 epochs in about 12 hours. I also didn't follow the structure of the paper for the generator (the upscaling of the dimensions of the image).